Renamed directory; bump up to 1.0.2
