+++ .gitignore
@@ -18,3 +18,6 @@ __pycache__
 # misc
 *_cache
 cspell.json
+
+# misc
+.DS_STORE
