+++ packages/core/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "jpyc_core_sdk"
-version = "1.0.1"
+name = "jpyc-core-sdk"
+version = "1.0.2"
 requires-python = ">=3.12"
 license = "MIT"
 description = "A Python SDK for interacting with JPYCv2 contracts"
@@ -26,12 +26,12 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 authors = [
-    { name = "mameta", email = "mameta.zk@gmail.com" },
     { name = "SeiyaKobayashi", email = "eng@seiya.work" },
+    { name = "mameta", email = "mameta.zk@gmail.com" },
 ]
 maintainers = [
-    { name = "mameta", email = "mameta.zk@gmail.com" },
     { name = "SeiyaKobayashi", email = "eng@seiya.work" },
+    { name = "mameta", email = "mameta.zk@gmail.com" },
 ]
 # Production dependencies
 dependencies = [
@@ -55,9 +55,14 @@ dev = [
     "ruff>=0.11.9",
 ]
 
+# setuptools
+
 [build-system]
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.data-files]
+artifacts = ["artifacts/*.json"]
 
 # pytest
