+++ tests/core/jpyc_methods/test_configure_minter.py
@@ -1,6 +1,6 @@
 import pytest
 
-from packages.core.src.utils.errors import (
+from packages.core.jpyc_core_sdk.utils.errors import (
     AccountNotInitialized,
     InvalidChecksumAddress,
     InvalidUint256,
